Move reset button from header bar into dataset controls.
