+++ src/components/HeaderBar.vue
@@ -56,54 +56,17 @@
         <v-tooltip location="bottom" activator="parent" :disabled="mobile" text="What's new"></v-tooltip>
         <v-icon>mdi-creation</v-icon>
       </v-btn>
-  
-      <icon-button
-        fa-icon="fa-arrows-rotate"
-        @activate="showConfirmReset = true"
-        tooltip-text="Reset app state"
-        :color="accentColor2"
-      >
-      </icon-button>
-      <v-dialog
-        v-model="showConfirmReset"
-        max-width="35%"
-      >
-        <v-card>
-          <v-card-text>
-            Are you sure you want to reset the app state? 
-            This will reset all of your regions, time ranges, 
-            and datasets, and cannot be undone.
-          </v-card-text>
-          <v-card-actions>
-            <v-btn
-              color="error"
-              @click="showConfirmReset = false"
-            >
-              No
-            </v-btn>
-            <v-btn
-              color="success"
-              @click="() => {
-                store.reset();
-                showConfirmReset = false;
-              }"
-            >
-              Yes
-            </v-btn>
-          </v-card-actions>
-        </v-card>
-      </v-dialog>
       <share-button
-          :source="currentUrl"
-          buttonColor="black"
-          :iconColor="accentColor2"
-          elevation="0"
-          size="small"
-          rounded="1"
-          :tooltip-disabled="mobile"
-          @click="shareButtonClickedCount += 1"
-          alert
-        />
+        :source="currentUrl"
+        buttonColor="black"
+        :iconColor="accentColor2"
+        elevation="0"
+        size="small"
+        rounded="1"
+        :tooltip-disabled="mobile"
+        @click="shareButtonClickedCount += 1"
+        alert
+      />
       <v-btn aria-role="menu" aria-label="Show menu" class="menu-button" variant="outlined" rounded="lg" :color="accentColor2" elevation="5">
         <v-icon size="x-large">mdi-menu</v-icon>
         <v-menu
@@ -199,7 +162,6 @@ const currentUrl = ref(window.location.href);
 const showChanges = ref(false);
 const showAboutData = ref(false);
 const showCredits = ref(false);
-const showConfirmReset = ref(false);
 
 const touchscreen = supportsTouchscreen();
